server-ex/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
uswgi

